--- a/ConditionalAccessExporter/ConditionalAccessExporter.csproj
+++ b/ConditionalAccessExporter/ConditionalAccessExporter.csproj
@@ -23,6 +23,7 @@
     <PackageReference Include="YamlDotNet" Version="13.7.1" />
     <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" />
+    <PackageReference Include="BenchmarkDotNet" Version="0.13.12" />
   </ItemGroup>
 
 </Project>