--- a/README.md
+++ b/README.md
@@ -35,6 +35,10 @@ ConditionalAccessExporter/
 - ✅ **Client Credential Authentication** - Uses Azure service principal for secure authentication
 - ✅ **Policy Export** - Retrieves all Conditional Access policies from Azure AD
 - ✅ **JSON Output** - Exports data in structured JSON format with metadata
+- ✅ **Performance Benchmarking** - Comprehensive performance monitoring and benchmarking suite
+- ✅ **Memory Usage Monitoring** - Track memory consumption and detect potential leaks
+- ✅ **Progress Indicators** - Real-time progress reporting for long-running operations
+- ✅ **Regression Testing** - Automated performance regression detection for CI/CD
 - ✅ **Comprehensive Error Handling** - Detailed error messages and permission guidance
 - ✅ **Security Best Practices** - Follows Azure security recommendations
 - ✅ **Documentation** - Complete documentation with examples and troubleshooting