… ConditionalAccessExporter/PerformanceBenchmarkProgram.cs
… ConditionalAccessExporter/Services/FileIOBenchmarks.cs
… ConditionalAccessExporter/Services/MemoryUsageBenchmarks.cs
… ConditionalAccessExporter/Services/PerformanceBenchmarkService.cs
… ConditionalAccessExporter/Services/PolicyProcessingBenchmarks.cs
… ConditionalAccessExporter/Utils/ProgressIndicator.cs
… PERFORMANCE_BENCHMARKING.md