# Performance Benchmarking and Monitoring Implementation

Resolves #93

## Overview

This PR implements comprehensive performance benchmarking and monitoring capabilities for CA_Scanner, addressing all requirements outlined in Issue #93.

## Key Features Implemented

### ✅ BenchmarkDotNet Integration
- Added BenchmarkDotNet package for professional-grade performance testing
- Created benchmarking infrastructure with proper configuration
- Implemented memory diagnostics and performance measurement

### ✅ Comprehensive Benchmark Categories

#### Policy Processing Benchmarks
- JSON serialization/deserialization performance
- Policy comparison operations
- Terraform conversion processes
- Policy validation benchmarks
- Parallel processing scenarios
- Scalability testing with varying dataset sizes

#### File I/O Benchmarks
- Synchronous vs asynchronous file operations
- Small vs large file processing performance
- Multiple file handling efficiency
- Streaming vs buffered operations
- Concurrent file processing

#### Memory Usage Benchmarks
- Object allocation patterns analysis
- Garbage collection impact measurement
- Memory leak detection capabilities
- Large dataset handling optimization
- Working set and memory delta tracking

### ✅ Performance Regression Testing
- Configurable performance thresholds
- Automated regression detection for CI/CD
- Performance quality gates
- Threshold violation alerts

### ✅ Memory Monitoring & Leak Detection
- Real-time memory usage tracking
- Pre/post garbage collection analysis
- Memory leak indicator calculations
- Working set monitoring
- Memory delta reporting

### ✅ Progress Indicators
- Real-time progress reporting for long operations
- ETA calculations and throughput metrics
- Memory usage during operation execution
- Customizable progress reporting interfaces
- Console-based progress visualization

### ✅ CLI Integration
- New `benchmark` command with comprehensive options
- Support for different benchmark types (all, benchmarks, regression, memory)
- Output file generation capabilities
- Help documentation and usage examples

## New Files Added

### Core Services
- `Services/PerformanceBenchmarkService.cs` - Main benchmarking orchestration
- `Services/PolicyProcessingBenchmarks.cs` - Policy operation benchmarks
- `Services/FileIOBenchmarks.cs` - File system performance tests
- `Services/MemoryUsageBenchmarks.cs` - Memory consumption analysis

### Utilities
- `Utils/ProgressIndicator.cs` - Progress monitoring and reporting
- `PerformanceBenchmarkProgram.cs` - Standalone benchmarking program

### Tests
- `Tests/PerformanceBenchmarkServiceTests.cs` - Service validation (17 tests)
- `Tests/ProgressIndicatorTests.cs` - Progress indicator testing (15 tests)

### Documentation
- `PERFORMANCE_BENCHMARKING.md` - Comprehensive usage guide
- Updated `README.md` with new features
- Enhanced `PERFORMANCE_OPTIMIZATIONS.md` with benchmarking references

## Usage Examples

### Running All Benchmarks
```bash
dotnet run --project ConditionalAccessExporter -- benchmark
```

### Running Specific Categories
```bash
# Performance benchmarks only
dotnet run --project ConditionalAccessExporter -- benchmark --type benchmarks

# Regression tests for CI/CD
dotnet run --project ConditionalAccessExporter -- benchmark --type regression

# Memory analysis
dotnet run --project ConditionalAccessExporter -- benchmark --type memory
```

### Saving Results
```bash
dotnet run --project ConditionalAccessExporter -- benchmark --output results.txt
```

## Sample Output

```
CA_Scanner Performance Benchmarking Tool
=========================================

Running performance regression tests...
✓ PolicyComparison: 109ms (Threshold: 5000ms) - PASS
✓ FileIO: 60ms (Threshold: 2000ms) - PASS

Policy Processing: 150/500 (30.0%) | Elapsed: 00:02:15 | ETA: 00:05:30 | Rate: 1.1 items/sec | Memory: 67.3 MB | Delta: +12.4 MB
```

## Testing

- **All 224 existing tests continue to pass**
- **32 new tests added** for benchmarking functionality
- **100% build success** with comprehensive error handling
- **Memory leak detection** validated and working
- **Performance regression tests** validated with configurable thresholds

## CI/CD Integration

The implementation supports seamless integration with CI/CD pipelines:

```yaml
- name: Run Performance Tests
  run: |
    dotnet run --project ConditionalAccessExporter -- benchmark --type regression
```

## Benefits

1. **Performance Insights** - Detailed analysis of bottlenecks and optimization opportunities
2. **Regression Detection** - Catch performance degradations early in development
3. **User Experience** - Progress indicators for long-running operations
4. **Optimization Guidance** - Data-driven performance improvement recommendations
5. **Quality Assurance** - Automated performance quality gates

## Repository Instructions Update

As requested in the issue comment, this implementation includes comprehensive documentation and follows established patterns to ensure consistent performance testing going forward. The `PERFORMANCE_BENCHMARKING.md` guide provides complete instructions for OpenHands and other contributors.

## Backward Compatibility

- All existing functionality remains unchanged
- No breaking changes to existing APIs
- New features are additive and optional
- Existing tests continue to pass without modification

## Performance Impact

- Benchmarking features are opt-in and don't affect normal operation
- Minimal memory overhead when not actively benchmarking
- Progress indicators use efficient timer-based updates
- All benchmark code follows async/await patterns

---

This implementation fully addresses Issue #93 requirements and provides a robust foundation for ongoing performance monitoring and optimization of the CA_Scanner application.